requirements.txt
alembic==1.14.1
annotated-types==0.7.0
appdirs==1.4.4
argcomplete==3.5.3
astroid==3.3.8
babel==2.17.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
dominate==2.9.1
Flask==3.1.0
flask-babel==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.10
importlib_resources==6.5.2
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pylint==3.3.4
python-barcode==0.15.1
python-dotenv==1.0.1
python-escpos==3.1
pytz==2025.1
pyusb==1.3.1
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
tenacity==9.0.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
